move app code to app/ directory
