+++ app/memory/base.py
