+++ app/prompt_templates/friendly_conversation.py
