+++ app/tools/web_access.py
@@ -1,5 +1,5 @@
-from settings import Settings
-from tools.base import BaseTool
+from app.settings import Settings
+from app.tools.base import BaseTool
 
 config = Settings.load_config()
